Remove async to simplify generator
Although not the only possibility, the use case I have in mind is for
the generator to handle one version at a time, and do nothing between
versions. So it is totally fine to just make each generator call the top
level of an async executor and still get the full benefits of async,
without making it required.